Add the mapping for DIII-D ec_launchers
This was primarily converted from OMAS by Claude Code.

The code.parameters entry that is used in OMAS to store the
DISPERSION data as toray.bhalf has been skipped for a few
reasons. 1) This looks like a hack in OMAS and it isn't clear
how important the data is. 2) The code parameters entries
break the logic in ods __contains__. That doesn't prevent the
fetch from being setup in imas_composer, but it would require
changes in omas for the testing to work which doesn't seem
essential at this time.

Some additional machinery to allow ec_launchers to be tested with
a different reference shot was also required because 200000 doesn't
have any EC heating.