Add mapping for the gas_injection
This was all essentially hardcoded in OMAS so now the definitions
come entirely from the yaml.

As long as these don't ever change on DIII-D this seems like a
reasonable implementation, but if they do it should be stored in
MDSplus instead. This should be confirmed with the ROs (Barret
Elward and Patrick Byrne according to Sharepoint).